requirements.txt
blessings==1.6
bpython==0.16
curtsies==0.2.11
greenlet==0.4.10
nose==1.3.7
Pygments==2.1.3
python-consul==0.6.1
requests==2.11.1
six==1.10.0
slack-consul==0.1
slacker==0.9.29
wcwidth==0.1.7
